MappingRelatedData: move to separate header and source files.
